+++ pymince/functional.py
@@ -1,4 +1,7 @@
 # -*- coding: utf-8 -*-
+import functools
+
+
 class classproperty:
     """
     Decorator that converts a method with a single cls argument into a property
@@ -20,3 +23,27 @@ class classproperty:
 
     def __get__(self, instance, cls=None):
         return self.fget(cls)
+
+
+def pipe(*fns):
+    """
+    Simple pipe function implementation using function composition.
+
+    :param fns: Functions to pipe.
+    :rtype: Callable[[Any], Any]
+
+    Examples:
+        from pymince.functional import pipe
+
+        addtwo = lambda n: n + 2
+        double = lambda n: n * 2
+        square = lambda n: n * n
+
+        fn = pipe(addtwo, double, square)
+        fn(1) # 36
+    """
+
+    def wrap(arg):
+        return functools.reduce(lambda v, fn: fn(v), fns, arg)
+
+    return wrap