+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - **Added:** Apply pre-commit badge to README.
 - **Added:** text.multireplace
 - **Added:** algorithm.sieve_of_eratosthenes. *Primes less than n.*
+- **Added:** functional.pipe. *Pipe function*
 - **Improved:** iterator: splitter, uniques, uniquer, consume, all_equals, all_identical, all_equal, has_only_one
 
 ## v2.0.0 (2022-12-22) ##